[ENTESB-14566] MailIntegrationTest may not receive expected number of messages
