+++ itests/standalone/basic/src/test/java/org/wildfly/camel/test/mail/MailIntegrationTest.java
@@ -21,16 +21,13 @@
 package org.wildfly.camel.test.mail;
 
 import java.io.File;
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
 import org.apache.camel.CamelContext;
 import org.apache.camel.ProducerTemplate;
-import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
-import org.apache.camel.impl.DefaultCamelContext;
-import org.apache.camel.support.jndi.JndiBeanRepository;
-import org.jboss.arquillian.container.test.api.Deployer;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
@@ -39,22 +36,28 @@ import org.jboss.as.arquillian.api.ServerSetupTask;
 import org.jboss.as.arquillian.container.ManagementClient;
 import org.jboss.dmr.ModelNode;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
+import org.jboss.shrinkwrap.api.asset.EmptyAsset;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
+import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.wildfly.camel.test.common.utils.DMRUtils;
+import org.wildfly.camel.test.mail.subA.MailSessionProducer;
 import org.wildfly.extension.camel.CamelAware;
+import org.wildfly.extension.camel.CamelContextRegistry;
 
 @CamelAware
-@ServerSetup({ MailIntegrationTest.MailSessionSetupTask.class })
 @RunWith(Arquillian.class)
+@ServerSetup({MailIntegrationTest.MailSessionSetupTask.class})
 public class MailIntegrationTest {
 
-    private static final String GREENMAIL_WAR = "greenmail.war";
-
+    static final Logger LOG = LoggerFactory.getLogger(MailIntegrationTest.class);
+    
     @ArquillianResource
-    Deployer deployer;
+    CamelContextRegistry contextRegistry;
 
     static class MailSessionSetupTask implements ServerSetupTask {
 
@@ -65,7 +68,7 @@ public class MailIntegrationTest {
                 .addStep("socket-binding-group=standard-sockets/remote-destination-outbound-socket-binding=mail-greenmail-smtp", "add(host=localhost, port=10025)")
                 .addStep("socket-binding-group=standard-sockets/remote-destination-outbound-socket-binding=mail-greenmail-pop3", "add(host=localhost, port=10110)")
                 .addStep("subsystem=mail/mail-session=greenmail", "add(jndi-name=java:jboss/mail/greenmail)")
-                .addStep("subsystem=mail/mail-session=greenmail/server=smtp", "add(outbound-socket-binding-ref=mail-greenmail-smtp, username=user1, password=password)")
+                .addStep("subsystem=mail/mail-session=greenmail/server=smtp", "add(outbound-socket-binding-ref=mail-greenmail-smtp, username=user1, password=password1)")
                 .addStep("subsystem=mail/mail-session=greenmail/server=pop3", "add(outbound-socket-binding-ref=mail-greenmail-pop3, username=user2, password=password2)")
                 .build();
 
@@ -87,51 +90,45 @@ public class MailIntegrationTest {
         }
     }
 
-    @Deployment
-    public static JavaArchive createDeployment() {
-        return ShrinkWrap.create(JavaArchive.class, "camel-mail-tests.jar");
-    }
-
-    @Deployment(managed = false, testable = false, name = GREENMAIL_WAR)
+    @Deployment(order = 1, testable = false, name = "greenmail")
     public static WebArchive createGreenmailDeployment() {
         return ShrinkWrap.createFromZipFile(WebArchive.class, new File("target/dependencies/greenmail-webapp.war"));
     }
 
+    @Deployment(order = 2)
+    public static JavaArchive createDeployment() throws IOException {
+        final JavaArchive archive = ShrinkWrap.create(JavaArchive.class, "camel-mail-cdi-tests.jar");
+        archive.addPackage(MailSessionProducer.class.getPackage());
+        archive.addAsManifestResource(EmptyAsset.INSTANCE, "beans.xml");
+        return archive;
+    }
+
     @Test
-    public void testMailEndpoint() throws Exception {
+    public void testMailEndpointWithCDIContext() throws Exception {
+    	
+        CamelContext camelctx = contextRegistry.getCamelContext("camel-mail-cdi-context");
+        Assert.assertNotNull("Camel context not null", camelctx);
 
-        CamelContext camelctx = new DefaultCamelContext(new JndiBeanRepository());
-        camelctx.addRoutes(new RouteBuilder() {
-            @Override
-            public void configure() throws Exception {
-                from("direct:start")
-                .to("smtp://localhost:10025?session=#java:jboss/mail/greenmail");
+        MockEndpoint mockEndpoint = camelctx.getEndpoint("mock:result", MockEndpoint.class);
+        mockEndpoint.setMinimumExpectedMessageCount(1);
 
-                from("pop3://user2@localhost:10110?delay=1000&session=#java:jboss/mail/greenmail&delete=true")
-                .to("mock:result");
-            }
-        });
+        Map<String, Object> mailHeaders = new HashMap<>();
+        mailHeaders.put("from", "user1@localhost");
+        mailHeaders.put("to", "user2@localhost");
+        mailHeaders.put("message", "Hello Kermit");
 
-        try {
-            deployer.deploy(GREENMAIL_WAR);
-
-            camelctx.start();
-
-            MockEndpoint mockEndpoint = camelctx.getEndpoint("mock:result", MockEndpoint.class);
-            mockEndpoint.setExpectedMessageCount(1);
-
-            Map<String, Object> mailHeaders = new HashMap<>();
-            mailHeaders.put("from", "user1@localhost");
-            mailHeaders.put("to", "user2@localhost");
-            mailHeaders.put("message", "Hello Kermit");
-
-            ProducerTemplate template = camelctx.createProducerTemplate();
+        ProducerTemplate template = camelctx.createProducerTemplate();
+        
+        long now = System.currentTimeMillis();
+        long end = now + 10000;
+        
+        while (now < end && mockEndpoint.getExchanges().isEmpty()) {
+            LOG.info("Sending: {}", mailHeaders);
             template.requestBodyAndHeaders("direct:start", null, mailHeaders);
-
-            mockEndpoint.assertIsSatisfied(5000);
-        } finally {
-            camelctx.close();
-            deployer.undeploy(GREENMAIL_WAR);
+            now = System.currentTimeMillis();
+            Thread.sleep(2000);
         }
+
+        mockEndpoint.assertIsSatisfied();
     }
 }
